optimize the gui
