+++ ScientificCalculator.py
@@ -1,6 +1,4 @@
-from PyQt5.QtGui import QIcon
-from PyQt5.QtWidgets import QApplication, QMainWindow, QWidget, QVBoxLayout, QLineEdit, QPushButton, QGridLayout
-import sys
+
 from typing import List
 from math import pi
 
@@ -231,108 +229,3 @@ class ScientificCalculator:
                 stack_cal.append(item)
 
         return stack_cal.pop()
-
-
-def add_superscript(text):
-    superscript_mapping = str.maketrans(
-        "0123456789+-=()abcdefghijklmnopqrstuvwxyzAB",
-        "⁰¹²³⁴⁵⁶⁷⁸⁹⁺⁻⁼⁽⁾ᵃᵇᶜᵈᵉᶠᵍʰⁱʲᵏˡᵐⁿᵒᵖᵠʳˢᵗᵘᵛʷˣʸᶻ½⅓"
-    )
-    return text.translate(superscript_mapping)
-
-
-class Calculator(QWidget):
-    def __init__(self):
-        super().__init__()
-        self.setWindowTitle('Scientific Calculator')
-
-        self.line_edit = QLineEdit(self)
-        self.line_edit.setReadOnly(True)
-
-        self.buttons = [
-            '7',  '8', '9', '/',
-            '4',  '5', '6', '*',
-            '1',  '2', '3', '-',
-            '0',  '.', '=', '+',
-            '!',  'sin', 'cos',
-            "e"+add_superscript("x"),
-            'x'+add_superscript("2"),
-            'x'+add_superscript("A"),
-            'x'+add_superscript("3"),
-            'x'+add_superscript("B")
-        ]
-
-        layout = QVBoxLayout()
-        layout.addWidget(self.line_edit)
-        
-        button_style = '''
-        QPushButton {
-            background-color: #4CAF50;
-            border: none;
-            color: white;
-            font-size: 18px;
-            height: 40px;
-        }
-
-        QPushButton:hover {
-            background-color: #45a049;
-        }
-        '''
-        
-        for button_text in self.buttons:
-            button = QPushButton(button_text)
-            button.setStyleSheet(button_style)
-            button.clicked.connect(self.button_clicked)
-            layout.addWidget(button)
-
-        layout = QGridLayout()
-        layout.addWidget(self.line_edit, 0, 0, 1, 4)
-
-        row = 1
-        col = 0
-        for button_text in self.buttons:
-            button = QPushButton(button_text)
-            button.setStyleSheet(button_style)
-            button.clicked.connect(self.button_clicked)
-            layout.addWidget(button, row, col)
-            col += 1
-            if col > 3:
-                col = 0
-                row += 1
-
-        self.setLayout(layout)
-
-    def button_clicked(self):
-        # 获取点击的按钮
-        button = self.sender()
-        text = button.text()
-
-        if text == '=':
-            try:
-                infix_expression = self.line_edit.text()
-                cal = ScientificCalculator()
-                res = cal.calculate_postfix(cal.postfix(cal.expression_pretreatment(infix_expression)))
-                self.line_edit.setText(str(res))
-            except:
-                self.line_edit.setText('Error')
-        else:
-            self.line_edit.setText(self.line_edit.text() + text)
-
-
-if __name__ == '__main__':
-    app = QApplication(sys.argv)
-    calculator = Calculator()
-    calculator.show()
-    sys.exit(app.exec_())
-
-
-# if __name__ == "__main__":
-#     cal = ScientificCalculator()
-#     while True:
-#         infix = input("Enter your calculate expression:\n")
-#         try:
-#             print(
-#                 f"Your result is {cal.calculate_postfix(cal.postfix(cal.expression_pretreatment(infix)))}\n"
-#             )
-#         except:
-#             print("ERROR\nEnter again!\n")
